Fixed a move_energy bug and added a tomo_xanes example script.
Previously, if calling TXM.move_energy() with the same value that the
beamline is currently set to, the epics PV would not ever return. Now
the function checks for this condition and returns immediately if if
the target energy matches the current energy.

The tomo_xanes.py script example was added to the documentation, it
may be converted to a stand-alone script in the future.